requirements.txt
flask==2.0.1
tensorflow==2.15.0
numpy==1.24.3
pillow==8.3.1
werkzeug==2.0.1
gunicorn==20.1.0
h5py==3.9.0
